requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bracex==2.6
cachetools==6.2.4
cbor2==5.8.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
claude-agent-sdk==0.1.18
click==8.3.1
cryptography==46.0.3
datasets==4.4.2
daytona==0.128.1
daytona_api_client==0.128.1
daytona_api_client_async==0.128.1
daytona_toolbox_api_client==0.128.1
daytona_toolbox_api_client_async==0.128.1
Deprecated==1.3.1
dill==0.4.0
dirhash==0.5.0
distro==1.9.0
dockerfile-parse==2.0.1
durationpy==0.10
e2b==2.9.0
environs==14.5.0
fastuuid==0.14.0
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.45.0
grpclib==0.4.9
h11==0.16.0
h2==4.3.0
harbor==0.1.32
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
linkify-it-py==2.0.3
litellm==1.80.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.2
mcp==1.25.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mini-swe-agent==1.8.1
modal==1.3.0.post1
mpmath==1.3.0
multidict==6.7.0
multipart==1.3.0
multiprocess==0.70.18
numpy==2.2.6
oauthlib==3.3.1
obstore==0.8.2
openai==1.99.9
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
pyaml==25.7.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
runloop_api_client==1.2.0
scantree==0.0.4
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sse-starlette==3.1.2
starlette==0.50.0
sympy==1.14.0
synchronicity==0.11.1
tenacity==9.1.2
textual==7.3.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tqdm==4.67.1
tree-sitter==0.25.2
typer==0.21.0
typer-slim==0.21.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.40.0
watchfiles==1.1.1
wcmatch==10.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
